Update call logic.

+++ src/main/java/dev/kovaliv/utils/DataUtils.java
@@ -3,7 +3,7 @@ package dev.kovaliv.utils;
 import java.util.Random;
 
 public class DataUtils {
-    public static long[] generateNumbers(int size) {
+    public static long[] generateNumbers(long size) {
         return new Random().longs(size).toArray();
     }
 }
